requirements.txt
requests==2.22.0
python-dateutil==2.8.2
discord.py==1.7.3
feedparser==6.0.8
scikit-image==0.18.3
pytz==2021.3
python-dotenv==0.19.1
numpy~=1.21.2
opencv-python-headless==4.5.3.56
asyncpg==0.24.0